+++ src/containerapp/azext_containerapp/tests/latest/test_containerapp_commands.py
@@ -716,6 +716,7 @@ class ContainerappServiceBindingTests(ScenarioTest):
         kafka_ca_name = 'kafka'
         mariadb_ca_name = 'mariadb'
         qdrant_ca_name = "qdrant"
+        milvus_ca_name = "milvus"
 
         create_containerapp_env(self, env_name, resource_group)
 
@@ -731,8 +732,16 @@ class ContainerappServiceBindingTests(ScenarioTest):
         self.cmd('containerapp service mariadb create -g {} -n {} --environment {}'.format(
             resource_group, mariadb_ca_name, env_name))
 
+
         self.cmd('containerapp service qdrant create -g {} -n {} --environment {}'.format(
             resource_group, qdrant_ca_name, env_name))
+
+        self.cmd('containerapp service milvus create -g {} -n {} --environment {}'.format(
+            resource_group, milvus_ca_name, env_name))
+
+        self.cmd('containerapp show -g {} -n {}'.format(resource_group, milvus_ca_name), checks=[
+            JMESPathCheck("properties.provisioningState", "Succeeded")
+        ])
 
         self.cmd('containerapp create -g {} -n {} --environment {} --image {} --bind postgres:postgres_binding redis'.format(
             resource_group, ca_name, env_name, image), checks=[
@@ -745,13 +754,15 @@ class ContainerappServiceBindingTests(ScenarioTest):
             JMESPathCheck('properties.template.serviceBinds[0].name', "redis"),
         ])
 
-        self.cmd('containerapp update -g {} -n {} --bind postgres:postgres_binding kafka mariadb qdrant'.format(
+        self.cmd('containerapp update -g {} -n {} --bind postgres:postgres_binding kafka mariadb qdrant milvus'.format(
             resource_group, ca_name, image), checks=[
+            JMESPathCheck("properties.provisioningState", "Succeeded"),
             JMESPathCheck('properties.template.serviceBinds[0].name', "redis"),
             JMESPathCheck('properties.template.serviceBinds[1].name', "postgres_binding"),
             JMESPathCheck('properties.template.serviceBinds[2].name', "kafka"),
             JMESPathCheck('properties.template.serviceBinds[3].name', "mariadb"),
-            JMESPathCheck('properties.template.serviceBinds[4].name', "qdrant")
+            JMESPathCheck('properties.template.serviceBinds[4].name', "qdrant"),
+            JMESPathCheck('properties.template.serviceBinds[5].name', "milvus")
         ])
 
         self.cmd('containerapp service postgres delete -g {} -n {} --yes'.format(
@@ -768,6 +779,9 @@ class ContainerappServiceBindingTests(ScenarioTest):
 
         self.cmd('containerapp service qdrant delete -g {} -n {} --yes'.format(
             resource_group, qdrant_ca_name, env_name))
+
+        self.cmd('containerapp service milvus delete -g {} -n {} --yes'.format(
+            resource_group, milvus_ca_name, env_name))
 
         self.cmd(f'containerapp delete -g {resource_group} -n {ca_name} --yes')
 
@@ -873,14 +887,14 @@ class ContainerappServiceBindingTests(ScenarioTest):
         ca_name = self.create_random_name(prefix='containerapp', length=24)
         mysqlserver = "mysqlflexsb"
         postgresqlserver = "postgresqlflexsb"
-
+        
         mysqlflex_json= self.cmd('mysql flexible-server create --resource-group {} --name {} --public-access {} -y'.format(resource_group, mysqlserver, "None")).output
         postgresqlflex_json= self.cmd('postgres flexible-server create --resource-group {} --name {} --public-access {} -y'.format(resource_group, postgresqlserver, "None")).output
         mysqlflex_dict = json.loads(mysqlflex_json)
 
         mysqlusername = mysqlflex_dict['username']
         mysqlpassword = mysqlflex_dict['password']
-
+        
         mysqldb = mysqlflex_dict['databaseName']
         flex_binding="mysqlflex_binding"
         postgresqlflex_dict = json.loads(postgresqlflex_json)